fix(web): consolidate Codex import in new session
